Deleted bulma framework
Even with bulma's modularity, it was outputting an incredible amount of unneeded/unused
CSS styles that did little but increase the final file's size. It's a great little
framework for rapid prototyping, but not one I'd want to use for a production site. The
other nuisance was with the unnecessary markup for styling purposes.

Some bulma elements were retained: mainly the variables split into three files in
sass/base/** along with the sass/utils/** files. Deleted unused and unnecessary
mixins, helpers, variables, etc. and added conditionals to some of the variables
used in utils/_mixins.sass even though those will most likely be heavily modifed or
even deleted outright in the future.